src-backend/rds-database-init/requirements.txt
psycopg2-binary==2.9.9
pymysql==1.1.0
